refactor: test 관련 파일 __tests__ 내 이동 및 테스트 코드 가독성 향상
