+++ __tests__/Cars.test.js
@@ -1,7 +1,7 @@
 import Car from '../src/Models/Car/Car.js';
 import createCars from '../src/Models/Cars/createCars.js';
 import { DuplicatedCarNameError } from '../src/Models/Cars/errors.js';
-import MoveStrategies from '../test/MoveStrategies.js';
+import MoveStrategies from './Fixtures/MoveStrategies.js';
 
 const { from, playOnce, getRecord } = createCars();
 describe('from() 테스트', () => {
@@ -20,7 +20,7 @@ describe('from() 테스트', () => {
             });
         });
 
-        describe('유효하면, 에러를 발생시키지 않는다.', () => {
+        describe('유효한 경우, 에러 없이 Car 배열을 생성한다.', () => {
             it.each([
                 { carNames: ['erica', 'Erica'] },
                 { carNames: ['gong0', 'Gong0', '1031', '1031!', '*****'] },
@@ -28,16 +28,7 @@ describe('from() 테스트', () => {
                 { carNames: ['!****', '*!***', '**!**', '***!*', '****!'] },
             ])('$carNames', ({ carNames }) => {
                 expect(() => from(carNames)).not.toThrow();
-            });
-        });
 
-        describe('중복이 없으면, Car 배열을 생성한다.', () => {
-            it.each([
-                { carNames: ['erica', 'Erica'] },
-                { carNames: ['gong0', 'Gong0', '1031', '1031!', '*****'] },
-                { carNames: ['*e*1C', '*e*1c', 'ERICA', 'Pan', 'theon'] },
-                { carNames: ['!****', '*!***', '**!**', '***!*', '****!'] },
-            ])('$carNames', ({ carNames }) => {
                 const cars = from(carNames);
 
                 cars.forEach((car, idx) => {
@@ -57,7 +48,8 @@ describe('playOnce() 테스트', () => {
     const cars = from(['erica', 'Erica', 'theon', 'yang', 'ryang']);
     playOnce(cars, new MoveStrategies('12345'));
 
-    it('Cars 배열 내 모든 자동차들이 올바르게 이동한다.', () => {
+    // TODO: 테스트 코드 가독성 개선 필요
+    it('Cars 배열 내 자동차들이 숫자가 4 이상이면 이동하고, 그렇지 않으면 이동하지 않는다.', () => {
         expect(cars.map((car) => car.getRecord().position)).toEqual([
             0, 0, 0, 1, 1,
         ]);
@@ -67,17 +59,38 @@ describe('playOnce() 테스트', () => {
 describe('getRecord() 테스트', () => {
     describe('Cars 내 모든 Car 정보를 반환한다.', () => {
         it.each([
-            { carNames: ['erica', 'Erica'] },
-            { carNames: ['gong0', 'Gong0', '1031', '1031!', '*****'] },
-            { carNames: ['*e*1C', '*e*1c', 'ERICA', 'Pan', 'theon'] },
-        ])('$carNames', ({ carNames }) => {
+            {
+                carNames: ['erica', 'Erica'],
+                expected: [
+                    { name: 'erica', position: 0 },
+                    { name: 'Erica', position: 0 },
+                ],
+            },
+            {
+                carNames: ['gong0', 'Gong0', '1031', '1031!', '*****'],
+                expected: [
+                    { name: 'gong0', position: 0 },
+                    { name: 'Gong0', position: 0 },
+                    { name: '1031', position: 0 },
+                    { name: '1031!', position: 0 },
+                    { name: '*****', position: 0 },
+                ],
+            },
+            {
+                carNames: ['*e*1C', '*e*1c', 'ERICA', 'Pan', 'theon'],
+                expected: [
+                    { name: '*e*1C', position: 0 },
+                    { name: '*e*1c', position: 0 },
+                    { name: 'ERICA', position: 0 },
+                    { name: 'Pan', position: 0 },
+                    { name: 'theon', position: 0 },
+                ],
+            },
+        ])('$carNames', ({ carNames, expected }) => {
             const cars = from(carNames);
             const roundRecord = getRecord(cars);
 
-            roundRecord.forEach((record, idx) => {
-                expect(record.name).toBe(carNames[idx]);
-                expect(record.position).toBe(0);
-            });
+            expect(roundRecord).toEqual(expected);
         });
     });
 });
