+++ __tests__/Fixtures/EmptyStrategy.js
@@ -1,4 +1,4 @@
-import MoveStrategy from '../src/Models/MoveStrategy/MoveStrategy.js';
+import MoveStrategy from '../../src/Models/MoveStrategy/MoveStrategy.js';
 
 export default class EmptyStrategy extends MoveStrategy {
     // generateNumber() {} 구현하지 않음
