+++ package.json
@@ -49,7 +49,11 @@
   "jest": {
     "transform": {
       "\\.[jt]sx?$": "babel-jest"
-    }
+    },
+    "testMatch": [
+      "**/__tests__/**/*.test.js", 
+      "**/?(*.)+(spec|test).[jt]s?(x)"
+    ]
   },
   "babel": {
     "presets": [
